backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
supabase==2.9.0
httpx==0.27.0
groq==0.4.2
requests==2.31.0

